Moved brickcolour.py and BrickColours.csv to a seperate folder
