+++ src/colour/brickcolour.py
